MasterRepository and AdminController added
