--- JobPortalAPI/Controller/EmployeeInfoController.cs
@@ -1,15 +1,0 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc;
-
-namespace JobPortalAPI.Controller
-{
-    [ApiController]
-    [Route("api/[controller]")]
-    public class EmployeeInfoController : ControllerBase
-    {
-        
-    }
-}
